% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data_moda}
\alias{get_data_moda}
\title{get_data_moda}
\usage{
get_data_moda(form_id, Token)
}
\arguments{
\item{form_id}{ID of users}

\item{Token}{Token of users}
}
\value{
dataframe
}
\description{
This function allows users to download data from Moda
}
\examples{
\dontrun{get_data_moda(56597, "9096b0028c4e68aa7a97122b889f3afe923b367b")}
}
