% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user.R
\name{get_user_moda}
\alias{get_user_moda}
\title{get_user_moda}
\usage{
get_user_moda(form_id, Token)
}
\arguments{
\item{form_id}{ID of users}

\item{Token}{Token of users}
}
\value{
dataframe
}
\description{
This function gives the list of users
}
\examples{
\dontrun{get_user_moda(56597, "9096b0028c4e68aa7a97122b889f3afe923b367b")}
}
